// formacion-web/src/interfaces/situaciones/i-situacion-string.ts
export interface ISituacionString {
    id: number;
    codApli: string;
    pro: number;
    gruGit: number;
    master: string;
    develop: string;
    actualizado: number;
    produccion: string;
    despl: string;
    was: number;
    maven: string;
    doc: string;
    pruebas: string;
    test: string;
    inf: string;
    terc: string;
}